Refatorar estrutura do aplicativo: remover GeminiService e reorganizar arquivos principais
